moved display/* to components/
