+++ qrshare/client/src/main/components/QR.svelte
