+++ qrshare/client/src/main/components/RouteListItem.svelte
