+++ qrshare/client/src/main/components/SearchListItem.svelte
