+++ qrshare/client/src/main/components/index.js
@@ -1,0 +1,5 @@
+export { default as QR } from "./QR.svelte";
+export { default as RouteList } from "./RouteList.svelte";
+export { default as RouteListItem } from "./RouteListItem.svelte";
+export { default as Searchbar } from "./Searchbar.svelte";
+export { default as SearchListItem } from "./SearchListItem.svelte";
