--- qrshare/client/src/main/display/index.js
@@ -1,4 +1,0 @@
-export { default as RouteList } from './RouteList.svelte'
-export { default as RouteListItem } from './RouteListItem.svelte'
-export { default as SearchListItem } from './SearchListItem.svelte'
-export { default as QR } from './QR.svelte'
